add proptypes
