+++ src/screens/Instructor/CourseSettings/Components/EmailList/index.js
@@ -2,15 +2,13 @@ import React, {useState} from 'react';
 import {
   CTFragment,
   CTForm, 
-  CTFormRow, 
-  CTInput,
-  CTFormHelp,
-  CTUploadButton
+  CTFormRow
 } from 'layout';
 import { Button } from 'pico-ui';
 import _ from 'lodash';
 import { user, uemail } from 'utils';
-import { EmailFilter } from './EmailFilter' 
+import EmailFilter from './EmailFilter';
+import UploadFile from './UploadFile';
 import './index.scss';
 
 
@@ -19,23 +17,6 @@ function EmailListWithRedux(props) {
   const [emails, setEmails] = useState([]);
   const [inputValue, setInputValue] = useState('');
   const [error, setError] = useState(null);
-
-  const onLoadFile = async (event) => {
-    let results = uemail.extract(event.target.result);
-    if (!Array.isArray(results)) return;
-    
-    setEmails(oldEmails => ([...oldEmails, ..._.difference(results, oldEmails)]));
-  };
-
-  const handleFileUpload = (files) => {
-    if (files.length > 0) {
-      for (let i = 0; i < files.length; i += 1) {
-        const reader = new FileReader()
-        reader.onload = onLoadFile;
-        reader.readAsText(files[i])
-      }
-    }
-  };
 
   return (
     <CTFragment className="email-list-container">
@@ -49,26 +30,13 @@ function EmailListWithRedux(props) {
       >     
         <CTFormRow>
           <CTFragment className="email-list-left">
-            <CTFormHelp title="INSTRUCTION">
-              <CTFragment>
-                Please upload a <strong>.csv/.txt file</strong> 
-                with a <strong>list of emails</strong>.
-              </CTFragment>
-              <hr />
-              <CTFragment className="email-list-uploadbtw-example">
-                <h4><strong>EXAMPLAE</strong></h4>
-                <CTFragment>{'<demo.txt>'}</CTFragment>
-                <CTFragment>shawn@university.edu</CTFragment>
-                <CTFragment>micheal2@university.edu</CTFragment>
-                <CTFragment>xiaoming@university.edu</CTFragment>
-                <CTFragment>...</CTFragment>
-              </CTFragment>
-            </CTFormHelp>
-            <CTUploadButton 
-              fluid 
-              id="upload-email-list"
-              onFileChange={handleFileUpload}
-              accept=".csv,.txt"
+            <UploadFile
+              emails={emails}
+              setEmails={setEmails}
+              inputValue={inputValue}
+              setInputValue={setInputValue}
+              error={error}
+              setError={setError}
             />
           </CTFragment>
